requirements.txt
Pillow>=10.0.0
PySide6>=6.6.0
numpy>=1.24.0
pytest>=7.4.0
